feat(seo): ItemList JSON-LD sur /star-players (wrapper serveur)
Termine le #4. Convertit /star-players en server component qui injecte
le JSON-LD CollectionPage + ItemList + BreadcrumbList (SSR), tout en
conservant l'UI interactive (déplacée dans StarPlayersClient.tsx).
Builder pur buildStarPlayersListSchema + 2 tests. Fetch non-bloquant
(safeServerJson) : la liste s'affiche même si l'API échoue.

Vérifié : typecheck OK, build OK (/star-players désormais SSR), 1226
tests web verts.
